+++ CHANGELOG.rst
@@ -5,6 +5,9 @@ Current
 -------
 
 - Add code of conduct
+- Add template subcommand to manage template
+- The template subcommand take and --add argument to store file in
+  the template folder, and add it into the manifest file
 
 0.0.1 (2017-07-15)
 ------------------